Move file reader/batcher to batchers package
