+++ cmd/helpers/extractorBuilder.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"rare/cmd/readProgress"
 	"rare/pkg/extractor"
+	"rare/pkg/extractor/batchers"
 	"rare/pkg/extractor/dirwalk"
 	"rare/pkg/logger"
 	"runtime"
@@ -73,7 +74,7 @@ func BuildExtractorFromArguments(c *cli.Context) *extractor.Extractor {
 		}
 		return ret
 	} else { // Read (no-follow) source file(s)
-		ret, err := extractor.New(openFilesToChan(dirwalk.GlobExpand(fileglobs, recursive), gunzip, concurrentReaders, batchSize), &config)
+		ret, err := extractor.New(batchers.OpenFilesToChan(dirwalk.GlobExpand(fileglobs, recursive), gunzip, concurrentReaders, batchSize), &config)
 		if err != nil {
 			logger.Fatalln(err)
 		}
