+++ cmd/helpers/readChannels_test.go
@@ -23,22 +23,3 @@ func TestTailLineToChan(t *testing.T) {
 
 	close(tailchan)
 }
-
-func TestOpenFilesToChan(t *testing.T) {
-	filenames := make(chan string, 5)
-	filenames <- "readChannels_test.go" // me!
-	close(filenames)
-
-	batches := openFilesToChan(filenames, false, 1, 1)
-
-	total := 0
-	var lastStart uint64 = 0
-	for batch := range batches {
-		assert.Greater(t, batch.BatchStart, lastStart)
-		lastStart = batch.BatchStart
-		total += len(batch.Batch)
-		assert.Equal(t, "readChannels_test.go", batch.Source)
-	}
-
-	assert.NotZero(t, total)
-}
