+++ pkg/extractor/batchers/common.go
@@ -1,0 +1,4 @@
+package batchers
+
+// ReadAheadBufferSize is the default size of the read-ahead buffer
+const ReadAheadBufferSize = 128 * 1024
