+++ pkg/extractor/batchers/tailBatcher.go
@@ -1,0 +1,1 @@
+package batchers
